+++ spec/proxy_app_spec.rb
@@ -147,6 +147,64 @@ RSpec.describe RedmineExtendedApi::ProxyApp do
       end
     end
 
+    context 'when requesting the roles show endpoint and the controller exposes accept_api_auth?' do
+      let(:env) { { 'PATH_INFO' => '/extended_api/roles/3.json' } }
+      let(:rewritten_env) { { 'PATH_INFO' => '/roles/3.json' } }
+      let(:request) { double('Request', path: '/roles/3.json', request_method: 'GET') }
+      let(:route_params) { { controller: 'roles', action: 'show', format: 'json', id: '3' } }
+      let(:response) { [200, { 'Content-Type' => 'application/json' }, ['{"role":{}}']] }
+
+      before do
+        roles_controller = Class.new do
+          class << self
+            def accept_api_auth?(action)
+              action.to_sym == :show
+            end
+          end
+        end
+
+        stub_const('RolesController', roles_controller)
+
+        allow(proxy_app).to receive(:rewrite_env).with(env).and_return(rewritten_env)
+        allow(proxy_app).to receive(:build_request).with(rewritten_env).and_return(request)
+        allow(routes).to receive(:recognize_path).with('/roles/3.json', method: 'GET').and_return(route_params)
+        allow(rails_app).to receive(:call).with(rewritten_env).and_return(response)
+      end
+
+      it 'proxies the request to Rails' do
+        expect(proxy_app.call(env)).to eq(response)
+      end
+    end
+
+    context 'when requesting the roles show endpoint and the controller exposes accept_api_auth_actions' do
+      let(:env) { { 'PATH_INFO' => '/extended_api/roles/3.json' } }
+      let(:rewritten_env) { { 'PATH_INFO' => '/roles/3.json' } }
+      let(:request) { double('Request', path: '/roles/3.json', request_method: 'GET') }
+      let(:route_params) { { controller: 'roles', action: 'show', format: 'json', id: '3' } }
+      let(:response) { [200, { 'Content-Type' => 'application/json' }, ['{"role":{}}']] }
+
+      before do
+        roles_controller = Class.new do
+          class << self
+            def accept_api_auth_actions
+              [:show]
+            end
+          end
+        end
+
+        stub_const('RolesController', roles_controller)
+
+        allow(proxy_app).to receive(:rewrite_env).with(env).and_return(rewritten_env)
+        allow(proxy_app).to receive(:build_request).with(rewritten_env).and_return(request)
+        allow(routes).to receive(:recognize_path).with('/roles/3.json', method: 'GET').and_return(route_params)
+        allow(rails_app).to receive(:call).with(rewritten_env).and_return(response)
+      end
+
+      it 'proxies the request to Rails' do
+        expect(proxy_app.call(env)).to eq(response)
+      end
+    end
+
     context 'when the request uses XML format' do
       let(:env) { { 'PATH_INFO' => '/extended_api/issues.xml' } }
       let(:rewritten_env) { { 'PATH_INFO' => '/issues.xml' } }